Rename `EventSubscriberInterface` into `EventSubscriber`
