+++ src/Event/Subscriber/MutantCreatingConsoleLoggerSubscriber.php
@@ -35,77 +35,51 @@ declare(strict_types=1);
 
 namespace Infection\Event\Listener;
 
-use Infection\Event\EventDispatcher\EventSubscriberInterface;
-use Infection\Event\InitialTestCaseCompleted;
-use Infection\Event\InitialTestSuiteFinished;
-use Infection\Event\InitialTestSuiteStarted;
-use Infection\TestFramework\TestFrameworkAdapter;
-use InvalidArgumentException;
+use Infection\Event\EventDispatcher\EventSubscriber;
+use Infection\Event\MutantCreated;
+use Infection\Event\MutantsCreatingFinished;
+use Infection\Event\MutantsCreatingStarted;
 use Symfony\Component\Console\Helper\ProgressBar;
 use Symfony\Component\Console\Output\OutputInterface;
 
 /**
  * @internal
  */
-final class InitialTestsConsoleLoggerSubscriber implements EventSubscriberInterface
+final class MutantCreatingConsoleLoggerSubscriber implements EventSubscriber
 {
     private $output;
     private $progressBar;
-    private $testFrameworkAdapter;
-    private $debug;
 
-    public function __construct(OutputInterface $output, TestFrameworkAdapter $testFrameworkAdapter, bool $debug)
+    public function __construct(OutputInterface $output)
     {
         $this->output = $output;
-        $this->testFrameworkAdapter = $testFrameworkAdapter;
-        $this->debug = $debug;
 
         $this->progressBar = new ProgressBar($this->output);
-        $this->progressBar->setFormat('verbose');
+        $this->progressBar->setFormat('Creating mutated files and processes: %current%/%max%');
     }
 
     public function getSubscribedEvents(): array
     {
         return [
-            InitialTestSuiteStarted::class => [$this, 'onInitialTestSuiteStarted'],
-            InitialTestSuiteFinished::class => [$this, 'onInitialTestSuiteFinished'],
-            InitialTestCaseCompleted::class => [$this, 'onInitialTestCaseCompleted'],
+            MutantsCreatingStarted::class => [$this, 'onMutantsCreatingStarted'],
+            MutantCreated::class => [$this, 'onMutantCreated'],
+            MutantsCreatingFinished::class => [$this, 'onMutantsCreatingFinished'],
         ];
     }
 
-    public function onInitialTestSuiteStarted(InitialTestSuiteStarted $event): void
+    public function onMutantsCreatingStarted(MutantsCreatingStarted $event): void
     {
-        try {
-            $version = $this->testFrameworkAdapter->getVersion();
-        } catch (InvalidArgumentException $e) {
-            $version = 'unknown';
-        }
-
-        $this->output->writeln([
-            '',
-            'Running initial test suite...',
-            '',
-            sprintf(
-                '%s version: %s',
-                $this->testFrameworkAdapter->getName(),
-                $version
-            ),
-            '',
-        ]);
-        $this->progressBar->start();
+        $this->output->writeln(['']);
+        $this->progressBar->start($event->getMutantCount());
     }
 
-    public function onInitialTestSuiteFinished(InitialTestSuiteFinished $event): void
-    {
-        $this->progressBar->finish();
-
-        if ($this->debug) {
-            $this->output->writeln(PHP_EOL . $event->getOutputText());
-        }
-    }
-
-    public function onInitialTestCaseCompleted(InitialTestCaseCompleted $event): void
+    public function onMutantCreated(MutantCreated $event): void
     {
         $this->progressBar->advance();
+    }
+
+    public function onMutantsCreatingFinished(MutantsCreatingFinished $event): void
+    {
+        $this->progressBar->finish();
     }
 }
