+++ tests/phpunit/AutoReview/ProjectCode/ProjectCodeProvider.php
@@ -50,8 +50,8 @@ use Infection\Console\OutputFormatter\OutputFormatter;
 use Infection\Console\OutputFormatter\ProgressFormatter;
 use Infection\Console\Util\PhpProcess;
 use Infection\Engine;
-use Infection\Event\Listener\MutantCreatingConsoleLoggerSubscriber;
-use Infection\Event\Listener\MutationGeneratingConsoleLoggerSubscriber;
+use Infection\Event\Subscriber\MutantCreatingConsoleLoggerSubscriber;
+use Infection\Event\Subscriber\MutationGeneratingConsoleLoggerSubscriber;
 use Infection\Finder\ComposerExecutableFinder;
 use Infection\Finder\FilterableFinder;
 use Infection\Finder\TestFrameworkFinder;
