+++ tests/phpunit/Event/Subscriber/CiInitialTestsConsoleLoggerSubscriberTest.php
