+++ tests/phpunit/Event/Subscriber/CiMutantCreatingConsoleLoggerSubscriberTest.php
