+++ tests/phpunit/Event/Subscriber/CiMutationGeneratingConsoleLoggerSubscriberTest.php
