+++ tests/phpunit/Event/Subscriber/InitialTestsConsoleLoggerSubscriberTest.php
